requirements.txt
aioredis==1.3.1
amqp==2.5.1
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
cffi==1.13.1
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-channels==0.7.0
django-cors-headers==2.5.3
django-filter==2.1.0
django-redis==4.11.0
django-timezone-field==4.0
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
flake8==3.5.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
kombu==4.6.5
lazy-object-proxy==1.4.3
Markdown==3.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==1.0.16
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-crontab==2.4.0
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.2
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
schedule==0.6.0
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
Unipath==1.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.0
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.6.0
